Refactor theme list modal to support extend
