+++ console/packages/shared/src/index.ts
@@ -9,3 +9,4 @@ export * from "./states/comment-subject-ref";
 export * from "./states/backup";
 export * from "./states/plugin-installation-tabs";
 export * from "./states/entity";
+export * from "./states/theme-list-tabs";
