+++ console/src/modules/interface/themes/constants/index.ts
@@ -1,0 +1,2 @@
+export const THEME_ALREADY_EXISTS_TYPE =
+  "https://halo.run/probs/theme-alreay-exists";
