+++ console/src/modules/interface/themes/types/index.ts
@@ -1,0 +1,10 @@
+export interface ThemeInstallationErrorResponse {
+  detail: string;
+  instance: string;
+  themeName: string;
+  requestId: string;
+  status: number;
+  timestamp: string;
+  title: string;
+  type: string;
+}
